requirements.txt
fastapi==0.92.0
jose==1.0.0
passlib==1.7.4
pydantic==1.8.2
python-dotenv==0.21.1
python_jose==3.3.0
SQLAlchemy==1.4.22
